removed unused imports and removed the local text parser from the Data File class
I moved the method that parses Strings back into an ArrayList of
expenses into the MainActivity because it was a centralized place that
could handle it. Instead of having each class contain a text parser, I
moved it into the MainActivity as a static method and made it smarter to
detect when Strings cannot be parsed into ArrayLists because it is now a
possibility.