Moved text parser into MainActivity. Created a dialog that allows users read encrypted expenses sent by other users.
Look at DataFile Commit to see the changes to text parser. The app now
has capabilities to send other users encrypted data that they can
decrypt and add to their own expenses. Upcoming there will be an
activity that allows users to temporarily view the data, but currently
it just adds the encrypted data to the users own list of expenses.