Replaced help text with toast
On the main activity and subsequent activities that followed the same strucutre and layout (Used the same layout), I have removed the text that says long press an item for options. Instead when the user presses an item a toast is displayed that asks the user to long press the item for options. This clears up the main screen and prevents unnecessary text from taking up screen real estate.